+++ CHANGES.md
@@ -42,12 +42,14 @@
 **Dependencies:**
 
 - babel-core: v6.24.1 → [v6.25.0](https://github.com/babel/babel/releases/tag/v6.25.0)
+- babel-plugin-transform-react-remove-prop-types: v0.4.5 → [v0.4.6](https://github.com/oliviertassinari/babel-plugin-transform-react-remove-prop-types/releases/tag/v0.4.6) - fix use in Node 4
 - babel-preset-es2017: v6.24.1
-- detect-port: v1.1.3 → [v1.2.0](https://github.com/node-modules/detect-port/releases/tag/1.2.0)
+- detect-port: v1.1.3 → [v1.2.1](https://github.com/node-modules/detect-port/releases)
 - extract-text-webpack-plugin: v2.1.0 → [v2.1.2](https://github.com/webpack-contrib/extract-text-webpack-plugin/blob/master/CHANGELOG.md#212-2017-06-08)
 - file-loader: v0.11.1 → [v0.11.2](https://github.com/webpack-contrib/file-loader/blob/master/CHANGELOG.md#0112-2017-06-05) - `publicPath` option changes
 - inquirer: v3.0.6 → [v3.1.0](https://github.com/SBoudrias/Inquirer.js/releases/tag/v3.1.0)
 - style-loader: v0.18.1 → [v0.18.2](https://github.com/webpack-contrib/style-loader/blob/master/CHANGELOG.md#0182-2017-06-05)
+- url-loader: v0.5.8 → [v0.5.9](https://github.com/webpack-contrib/url-loader/releases/tag/v0.5.9)
 - [webpack-dev-server](https://github.com/webpack/webpack-dev-server/): v2.4.5
 
 # 0.16.3 / 2017-05-31